Java GUI project
The aim of this assignment was to gain experience with oriented
programming principles of inheritance and polymorphism.

The user can create a new shape by clicking anywhere
within the panel area of the program. The properties of
the newly created shape are based on the current values
saved in the appropriate UI fields (e.g. height, width etc). They can
click on a shape to change its height, width, colour and border
thickness and well as its bouncing path.